Custom_AI/requirements.txt
fastapi
uvicorn
openai>=1.6.0
requests
pydantic
python-docx
pypdf
openpyxl
python-pptx
pillow
